Update viewer UI
