+++ nerfstudio/viewer/app/src/index.scss
@@ -175,8 +175,9 @@ $sidebar-width: 100% !default;
 
 .CameraPanel-slider-container {
   margin-top: 40px;
-  margin-left: 20px;
-  margin-right: 20px;
+  margin-left: auto;
+  margin-right: auto;
+  width: 70%;
 }
 
 .CameraPanel-slider-button-container {
