+++ nerfstudio/viewer/app/src/modules/SidePanel/CameraPanel/CameraPanel.jsx
@@ -15,12 +15,15 @@ import { MeshLine, MeshLineMaterial } from 'meshline';
 import { useContext, useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import AddAPhotoIcon from '@mui/icons-material/AddAPhoto';
 import ArrowBackIosNewIcon from '@mui/icons-material/ArrowBackIosNew';
 import ArrowForwardIosIcon from '@mui/icons-material/ArrowForwardIos';
 import { CSS2DObject } from 'three/examples/jsm/renderers/CSS2DRenderer';
 import ContentPasteGoIcon from '@mui/icons-material/ContentPasteGo';
 import DeleteIcon from '@mui/icons-material/Delete';
 import FirstPageIcon from '@mui/icons-material/FirstPage';
+import FileDownloadOutlinedIcon from '@mui/icons-material/FileDownloadOutlined';
+import FileUploadOutlinedIcon from '@mui/icons-material/FileUploadOutlined';
 import IconButton from '@mui/material/IconButton';
 import LastPageIcon from '@mui/icons-material/LastPage';
 import PauseIcon from '@mui/icons-material/Pause';
@@ -562,58 +565,16 @@ export default function CameraPanel(props) {
     });
   };
 
-  const setUp = () => {
-    const rot = camera_main.rotation;
-    const unitY = new THREE.Vector3(0, 1, 0);
-    const upVec = unitY.applyEuler(rot);
-
-    const grid = sceneTree.find_object_no_create(['Grid']);
-    grid.setRotationFromEuler(rot);
-
-    const pos = new THREE.Vector3();
-    camera_main.getWorldPosition(pos);
-    camera_main.up.set(upVec.x, upVec.y, upVec.z);
-    sceneTree.metadata.camera_controls.updateCameraUp();
-    sceneTree.metadata.camera_controls.setLookAt(pos.x, pos.y, pos.z, 0, 0, 0);
-    const points = [new THREE.Vector3(0, 0, 0), upVec.multiplyScalar(2)];
-    const geometry = new THREE.BufferGeometry().setFromPoints(points);
-    const material = new THREE.LineBasicMaterial({
-      color: 0xaa46fc,
-      linewidth: 1,
-    });
-    const line = new THREE.LineSegments(geometry, material);
-    sceneTree.set_object_from_path(['Viewer Up Vector'], line);
-  };
-
   return (
     <div className="CameraPanel">
-      <div className="CameraPanel-top-button">
-        <Button size="small" variant="outlined" onClick={setUp}>
-          Reset Up Direction
-        </Button>
-      </div>
       <div>
-        <div className="CameraPanel-top-button">
-          <Button size="small" variant="outlined" onClick={add_camera}>
-            Add Camera
-          </Button>
-        </div>
-        <div className="CameraPanel-top-button">
-          <Button
-            size="small"
-            className="CameraPanel-top-button"
-            variant="outlined"
-            onClick={export_camera_path}
-          >
-            Export Path
-          </Button>
-        </div>
         <div className="CameraPanel-top-button">
           <Button
             size="small"
             className="CameraPanel-top-button"
             component="label"
             variant="outlined"
+            startIcon={<FileUploadOutlinedIcon />}
           >
             Load Path
             <input
@@ -626,6 +587,17 @@ export default function CameraPanel(props) {
           </Button>
         </div>
         <div className="CameraPanel-top-button">
+          <Button
+            size="small"
+            className="CameraPanel-top-button"
+            variant="outlined"
+            startIcon={<FileDownloadOutlinedIcon />}
+            onClick={export_camera_path}
+          >
+            Export Path
+          </Button>
+        </div>
+        <div className="CameraPanel-top-button">
           <Tooltip title="Copy Cmd to Clipboard">
             <IconButton onClick={copy_cmd_to_clipboard}>
               <ContentPasteGoIcon />
@@ -633,7 +605,147 @@ export default function CameraPanel(props) {
           </Tooltip>
         </div>
       </div>
+      <div className="CameraList-row-time-interval">
+        <TextField
+          label="Height"
+          inputProps={{
+            inputMode: 'numeric',
+            pattern: '[+-]?([0-9]*[.])?[0-9]+',
+          }}
+          size="small"
+          onChange={(e) => {
+            if (e.target.validity.valid) {
+              setUIRenderHeight(e.target.value);
+            }
+          }}
+          onBlur={(e) => {
+            if (e.target.validity.valid) {
+              if (e.target.value !== '') {
+                setRenderHeight(parseInt(e.target.value, 10));
+              } else {
+                setUIRenderHeight(render_height);
+              }
+            }
+          }}
+          value={ui_render_height}
+          error={ui_render_height <= 0}
+          helperText={ui_render_height <= 0 ? 'Required' : ''}
+          variant="standard"
+        />
+        <TextField
+          label="Width"
+          inputProps={{
+            inputMode: 'numeric',
+            pattern: '[+-]?([0-9]*[.])?[0-9]+',
+          }}
+          size="small"
+          onChange={(e) => {
+            if (e.target.validity.valid) {
+              setUIRenderWidth(e.target.value);
+            }
+          }}
+          onBlur={(e) => {
+            if (e.target.validity.valid) {
+              if (e.target.value !== '') {
+                setRenderWidth(parseInt(e.target.value, 10));
+              } else {
+                setUIRenderWidth(render_width);
+              }
+            }
+          }}
+          value={ui_render_width}
+          error={ui_render_width <= 0}
+          helperText={ui_render_width <= 0 ? 'Required' : ''}
+          variant="standard"
+        />
+        <TextField
+          label="FOV"
+          inputProps={{
+            inputMode: 'numeric',
+            pattern: '[+-]?([0-9]*[.])?[0-9]+',
+          }}
+          onChange={(e) => {
+            if (e.target.validity.valid) {
+              setUIFieldOfView(e.target.value);
+            }
+          }}
+          onBlur={(e) => {
+            if (e.target.validity.valid) {
+              if (e.target.value !== '') {
+                setFOV(parseInt(e.target.value, 10));
+              } else {
+                setUIFieldOfView(field_of_view);
+              }
+            }
+          }}
+          value={ui_field_of_view}
+          error={ui_field_of_view <= 0}
+          helperText={ui_field_of_view <= 0 ? 'Required' : ''}
+          variant="standard"
+        />
+      </div>
+      <div className="CameraList-row-time-interval">
+        <TextField
+          label="Seconds"
+          inputProps={{
+            inputMode: 'numeric',
+            pattern: '[+-]?([0-9]*[.])?[0-9]+',
+          }}
+          size="small"
+          onChange={(e) => {
+            if (e.target.validity.valid) {
+              setUISeconds(e.target.value);
+            }
+          }}
+          onBlur={(e) => {
+            if (e.target.validity.valid) {
+              if (e.target.value !== '') {
+                setSeconds(parseInt(e.target.value, 10));
+              } else {
+                setUISeconds(seconds);
+              }
+            }
+          }}
+          value={ui_seconds}
+          error={ui_seconds <= 0}
+          helperText={ui_seconds <= 0 ? 'Required' : ''}
+          variant="standard"
+        />
+        <TextField
+          label="FPS"
+          inputProps={{ inputMode: 'numeric', pattern: '[0-9]*' }}
+          size="small"
+          onChange={(e) => {
+            if (e.target.validity.valid) {
+              setUIfps(e.target.value);
+            }
+          }}
+          onBlur={(e) => {
+            if (e.target.validity.valid) {
+              if (e.target.value !== '') {
+                setFps(parseInt(e.target.value, 10));
+              } else {
+                setUIfps(fps);
+              }
+            }
+          }}
+          value={ui_fps}
+          error={ui_fps <= 0}
+          helperText={ui_fps <= 0 ? 'Required' : ''}
+          variant="standard"
+        />
+      </div>
       <div>
+        <div className="CameraPanel-top-button">
+          <Button
+            size="small"
+            variant="outlined"
+            startIcon={<AddAPhotoIcon />}
+            onClick={add_camera}
+          >
+            Add Camera
+          </Button>
+        </div>
         <div className="CameraPanel-top-button">
           <Tooltip className="curve-button" title="Close/Open camera spline">
             {!is_cycle ? (
@@ -790,136 +902,6 @@ export default function CameraPanel(props) {
         >
           <LastPageIcon />
         </Button>
-      </div>
-      <div className="CameraList-row-time-interval">
-        <TextField
-          label="Height"
-          inputProps={{
-            inputMode: 'numeric',
-            pattern: '[+-]?([0-9]*[.])?[0-9]+',
-          }}
-          size="small"
-          onChange={(e) => {
-            if (e.target.validity.valid) {
-              setUIRenderHeight(e.target.value);
-            }
-          }}
-          onBlur={(e) => {
-            if (e.target.validity.valid) {
-              if (e.target.value !== '') {
-                setRenderHeight(parseInt(e.target.value, 10));
-              } else {
-                setUIRenderHeight(render_height);
-              }
-            }
-          }}
-          value={ui_render_height}
-          error={ui_render_height <= 0}
-          helperText={ui_render_height <= 0 ? 'Required' : ''}
-          variant="standard"
-        />
-        <TextField
-          label="Width"
-          inputProps={{
-            inputMode: 'numeric',
-            pattern: '[+-]?([0-9]*[.])?[0-9]+',
-          }}
-          size="small"
-          onChange={(e) => {
-            if (e.target.validity.valid) {
-              setUIRenderWidth(e.target.value);
-            }
-          }}
-          onBlur={(e) => {
-            if (e.target.validity.valid) {
-              if (e.target.value !== '') {
-                setRenderWidth(parseInt(e.target.value, 10));
-              } else {
-                setUIRenderWidth(render_width);
-              }
-            }
-          }}
-          value={ui_render_width}
-          error={ui_render_width <= 0}
-          helperText={ui_render_width <= 0 ? 'Required' : ''}
-          variant="standard"
-        />
-        <TextField
-          label="FOV"
-          inputProps={{
-            inputMode: 'numeric',
-            pattern: '[+-]?([0-9]*[.])?[0-9]+',
-          }}
-          onChange={(e) => {
-            if (e.target.validity.valid) {
-              setUIFieldOfView(e.target.value);
-            }
-          }}
-          onBlur={(e) => {
-            if (e.target.validity.valid) {
-              if (e.target.value !== '') {
-                setFOV(parseInt(e.target.value, 10));
-              } else {
-                setUIFieldOfView(field_of_view);
-              }
-            }
-          }}
-          value={ui_field_of_view}
-          error={ui_field_of_view <= 0}
-          helperText={ui_field_of_view <= 0 ? 'Required' : ''}
-          variant="standard"
-        />
-      </div>
-      <div className="CameraList-row-time-interval">
-        <TextField
-          label="Seconds"
-          inputProps={{
-            inputMode: 'numeric',
-            pattern: '[+-]?([0-9]*[.])?[0-9]+',
-          }}
-          size="small"
-          onChange={(e) => {
-            if (e.target.validity.valid) {
-              setUISeconds(e.target.value);
-            }
-          }}
-          onBlur={(e) => {
-            if (e.target.validity.valid) {
-              if (e.target.value !== '') {
-                setSeconds(parseInt(e.target.value, 10));
-              } else {
-                setUISeconds(seconds);
-              }
-            }
-          }}
-          value={ui_seconds}
-          error={ui_seconds <= 0}
-          helperText={ui_seconds <= 0 ? 'Required' : ''}
-          variant="standard"
-        />
-        <TextField
-          label="FPS"
-          inputProps={{ inputMode: 'numeric', pattern: '[0-9]*' }}
-          size="small"
-          onChange={(e) => {
-            if (e.target.validity.valid) {
-              setUIfps(e.target.value);
-            }
-          }}
-          onBlur={(e) => {
-            if (e.target.validity.valid) {
-              if (e.target.value !== '') {
-                setFps(parseInt(e.target.value, 10));
-              } else {
-                setUIfps(fps);
-              }
-            }
-          }}
-          value={ui_fps}
-          error={ui_fps <= 0}
-          helperText={ui_fps <= 0 ? 'Required' : ''}
-          variant="standard"
-        />
       </div>
       <div className="CameraList-container">
         <CameraList
