+++ nerfstudio/viewer/app/src/modules/SidePanel/ScenePanel/index.jsx
@@ -1,0 +1,3 @@
+import ScenePanel from './ScenePanel';
+
+export default ScenePanel;
