+++ nerfstudio/viewer/app/src/themes/theme.ts
@@ -123,5 +123,13 @@ export const appTheme = createTheme({
         },
       },
     },
+    MuiList: {
+      styleOverrides: {
+        padding: {
+          paddingTop: 4,
+          paddingBottom: 4,
+        },
+      },
+    },
   },
 });
